+++ friends/src/utils/withAuth.js
@@ -1,0 +1,11 @@
+import axios from 'axios';
+export const withAuth = () => {
+  const token = localStorage.getItem('token');
+  return axios.create({
+    baseURL: 'http://localhost:5000/api',
+    headers: {
+      'Content-Type': 'application/json',
+      Authorization: token
+    }
+  });
+};
